페이징 처리중...

+++ resonos/SQL/resonos/프로시저.sql
@@ -325,7 +325,7 @@ BEGIN
 
     CREATE TABLE IF NOT EXISTS `track_review` (
         `id` BIGINT NOT NULL,
-        `rating` FLOAT NOT NULL,
+        `rating` INT NOT NULL,
         `content` TEXT NOT NULL,
         `blinded` BOOLEAN NOT NULL,
         `likes` INT NOT NULL DEFAULT 0,
